requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.9.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.12.0
bleach==5.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
Cython==0.29.32
databricks-cli==0.17.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.1
dpath==2.1.4
dulwich==0.20.50
dvc==2.40.0
dvc-data==0.28.5
dvc-http==2.30.0
dvc-objects==0.14.0
dvc-render==0.0.17
dvc-studio-client==0.3.0
dvc-task==0.1.9
dvclive==1.3.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grandalf==0.6
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
iterative-telemetry==0.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==5.2.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mlflow==2.1.1
mmcv-full==1.7.0
-e git+https://github.com/open-mmlab/mmdetection.git@31c84958f54287a8be2b99cbf87a6dcf12e57753#egg=mmdet
mmpycocotools==12.0.3
model-index==0.1.11
multidict==6.0.4
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.21.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
openmim==0.3.3
ordered-set==4.1.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.9.3
scmrepo==0.1.5
Send2Trash==1.8.0
shap==0.41.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
yapf==0.32.0
yarl==1.8.2
zc.lockfile==2.0
zipp==3.11.0
